+++ acceptance/tests/bolt_apply.rb
@@ -10,8 +10,8 @@ test_name "bolt apply should apply manifest block on remote hosts via ssh and wi
   winrm_nodes = select_hosts(roles: ['winrm'])
   targets = ssh_nodes + winrm_nodes
 
-  # Puppet 6 doesn't support OSX 10.11, so skip those hosts if present
-  targets -= select_hosts(platform: [/osx-10.11/, /fedora-31/])
+  # the puppet_agent::install task does not support osx1015 yet, so skip those hosts if present
+  targets -= select_hosts(platform: [/osx-10.15/])
 
   skip_test('no applicable nodes to test on') if targets.empty?