+++ CHANGELOG.md
@@ -13,7 +13,7 @@
   The `puppetfile install` command now supports a `--puppetfile` option that accepts a relative or absolute path
   to a Puppetfile.
 
-* **Update reboot plan parameter `nodes` to `targets`** ([puppetlabs-reboot#223](https://github.com/puppetlabs/reboot/pulls/223)
+* **Update reboot plan parameter `nodes` to `targets`** ([puppetlabs-reboot#223](https://github.com/puppetlabs/puppetlabs-reboot/pull/223))
 
   Users who explicitly set the `nodes` parameter will need to update the parameter name to
   `targets`. Calling the `reboot` plan with `-t` or `run_plan('reboot', $mytargets)` behaves the same